==> Comprehensive Machine Learning projects/SpaceX_Landing_Prediction/3_Data_Wrangling.ipynb <==
# **Space X  Falcon 9 First Stage Landing Prediction**

## Data wrangling


In this part of the project, we will perform some Exploratory Data Analysis (EDA) to find some patterns in the data and determine what would be the label for training supervised models.

In this notebook we will mainly convert the different outcomes into Training Labels with `1` means the booster successfully landed `0` means it was unsuccessful.

## Objectives

Perform exploratory  Data Analysis and determine Training Labels

*   Exploratory Data Analysis
*   Determine Training Labels

***

## Import Libraries and Define Auxiliary Functions

We will import the following libraries.

import pandas as pd
import numpy as np

### Data Analysis

Load Space X dataset, from last section.

df = pd.read_csv('dataset_part_1.csv')
df.head(10)

Identify and calculate the percentage of the missing values in each attribute

df.isnull().sum()/df.count()*100

Identify which columns are numerical and categorical:

df.dtypes

### Calculating the number of launches on each site

The data contains several Space X  launch facilities: <a href='https://en.wikipedia.org/wiki/List_of_Cape_Canaveral_and_Merritt_Island_launch_sites?utm_medium=Exinfluencer&utm_source=Exinfluencer&utm_content=000026UJ&utm_term=10006555&utm_id=NA-SkillsNetwork-Channel-SkillsNetworkCoursesIBMDS0321ENSkillsNetwork26802033-2021-01-01'>Cape Canaveral Space</a> Launch Complex 40  <b>VAFB SLC 4E </b> , Vandenberg Air Force Base Space Launch Complex 4E <b>(SLC-4E)</b>, Kennedy Space Center Launch Complex 39A <b>KSC LC 39A </b>.The location of each Launch Is placed in the column <code>LaunchSite</code>


Next, let's see the number of launches for each site.


df['LaunchSite'].value_counts()

Each launch aims to an dedicated orbit, and here are some common orbit types:

* **LEO**: Low Earth orbit 
* **VLEO**: Very Low Earth orbits 
* **GTO**: A geosyncronous orbit 
* **SSO (or SO)**: A Sun-synchronous orbit
* **ES-L1**: At Lagrange points
* **HEO**: A highly elliptical orbit
* **ISS**: International space station
* **MEO**: Geocentric orbits above LEO and lower than GTO
* **HEO**: Geocentric orbits above GTO
* **GEO**: A circular geosyncronous orbit
* **PO**: An orbit in which a satellite passes above both poles of the body being orbited

### Calculating the number and occurrence of each orbit

df['Orbit'].value_counts()

### Calculating the number and occurence of mission outcome per orbit type

landing_outcomes = df['Outcome'].value_counts()

<code>True Ocean</code> means the mission outcome was successfully  landed to a specific region of the ocean while <code>False Ocean</code> means the mission outcome was unsuccessfully landed to a specific region of the ocean. <code>True RTLS</code> means the mission outcome was successfully  landed to a ground pad <code>False RTLS</code> means the mission outcome was unsuccessfully landed to a ground pad.<code>True ASDS</code> means the mission outcome was successfully  landed to a drone ship <code>False ASDS</code> means the mission outcome was unsuccessfully landed to a drone ship. <code>None ASDS</code> and <code>None None</code> these represent a failure to land.


for i, outcome in enumerate(landing_outcomes.keys()):
    print(i, outcome)

We create a set of outcomes where the second stage did not land successfully:

bad_outcomes = set(landing_outcomes.keys()[[1,3,5,6,7]])
bad_outcomes

### Creating a landing outcome label from Outcome column

# landing_class = 0 if bad_outcome
# landing_class = 1 otherwise
outcome_list = df['Outcome'].tolist()
landing_class = [0 if x in bad_outcomes else 1 for x in outcome_list]

This variable will represent the classification variable that represents the outcome of each launch. If the value is zero, the  first stage did not land successfully; one means  the first stage landed Successfully

df['Class'] = landing_class
df[['Class']].head(8)

df.head(5)

We can use the following line of code to determine  the success rate:

df['Class'].mean()

We can now export it to a CSV for the next section.

df.to_csv("dataset_part_2.csv", index=False)
