==> Comprehensive Machine Learning projects/SpaceX_Landing_Prediction/4_EDA_with_SQL.ipynb <==
# **Space X  Falcon 9 First Stage Landing Prediction**

## Exploratory Data Analysis with SQL

In this part of the project, we will perform some Exploratory Data Analysis (EDA) using SQL by querying data stored in a PostgreSQL database. 

This analysis will give us additional understanding of the data.

## Objectives

In this Python notebook we will:

1.  Understand the Spacex DataSet
2.  Load the dataset from the corresponding table in a PostgreSQL database
3.  Execute SQL queries to answer questions that guide our analysis

***

First let's import required packages for this lab

import psycopg2
from config import config
from sqlalchemy.engine.url import URL
%load_ext sql

The SpaceX dataset has been loaded into a local PostgreSQL using the csv output ('dataset_part_2.csv') from part 3 of the project: 'Data Wrangling'.

We will be connecting to the local database server using a custom config function.

## Connecting to PostgreSQL

# Read connection parameters
params = config()
# Create the Connect string
url = URL.create(**params)
print(url)

Sensitive code cells have been removed.

Let's check our connection by looking at a few rows from the table

%sql select * from spacextbl limit 5;

## Data Exploration

##### Let's take a look at the names of the unique launch sites  in the space mission

%sql select DISTINCT (launch_site) from spacextbl;

##### Next, let's display launch sites where launch sites begin with the string 'CCA'

%sql select * from SPACEXTBL where launch_site like 'CCA%' limit 5

##### What is the total payload mass carried by boosters launched by NASA (CRS)?

%sql select sum(payload_mass_kg) as Total from spacextbl where customer = 'NASA (CRS)'

##### What is the average payload mass carried by booster version F9 v1.1?

%sql select avg(payload_mass_kg) as Average_Payload from spacextbl where booster_version like 'F9 v1.1%'

##### On what date was the first successful landing outcome on a ground pad was acheived?

%sql select min(date_) from spacextbl where landing_outcome = 'Success (ground pad)'

##### Has SpaceX carried out any launches on a Friday?

%sql select * from spacextbl where EXTRACT(DOW from date_) = 5 limit 5

##### What are the names of the boosters which have success in drone ship and have payload mass greater than 4000 but less than 6000?

%sql select booster_version from spacextbl where landing_outcome = 'Success (drone ship)' and (payload_mass_kg between 4000 and 6000)

##### What are the total number of successful and failure mission outcomes?

%sql select mission_outcome, COUNT(*) as Total from spacextbl group by mission_outcome

##### What are the names of the booster_versions which have carried the maximum payload mass?

# Using a subquery
%sql select Distinct (booster_version) from spacextbl where payload_mass_kg = (select max(payload_mass_kg) from spacextbl)

##### Let's see the failed landing_outcomes on drone ships, their booster versions, and launch site names in 2015.

%sql select booster_version, launch_site from spacextbl where landing_outcome = 'Failure (drone ship)' and EXTRACT(YEAR from date_) = '2015'

##### Lastly, let us rank the count of landing outcomes (such as Failure (drone ship) or Success (ground pad)) between the date 2010-06-04 and 2017-03-20, in descending order.

%sql select landing_outcome, count(*) as Count from spacextbl where date_ between '2010-06-04' and '2017-03-20' group by landing_outcome order by Count DESC
